Refactor load tasklist
